requirements.txt

# libraries need to be installed
streamlit
streamlit_extras
google-generativeai
python-dotenv
googletrans==4.0.0-rc1
Pillow
os


